@@ -2,7 +2,7 @@ name: google_sign_in
 description: Flutter plugin for Google Sign-In, a secure authentication system
   for signing in with a Google account on Android and iOS.
 homepage: https://github.com/flutter/plugins/tree/master/packages/google_sign_in/google_sign_in
-version: 4.5.9
+version: 5.0.0-nullsafety
 
 flutter:
   plugin:
@@ -16,16 +16,10 @@ flutter:
         default_package: google_sign_in_web
 
 dependencies:
-  google_sign_in_platform_interface: ^1.1.1
+  google_sign_in_platform_interface: ^2.0.0-nullsafety
   flutter:
     sdk: flutter
-  meta: ^1.0.4
-  # The design on https://flutter.dev/go/federated-plugins was to leave
-  # this constraint as "any". We cannot do it right now as it fails pub publish
-  # validation, so we set a ^ constraint.
-  # TODO(amirh): Revisit this (either update this part in the  design or the pub tool).
-  # https://github.com/flutter/flutter/issues/46264
-  google_sign_in_web: ^0.9.1
+  meta: ^1.3.0-nullsafety.6
 
 dev_dependencies:
   http: ^0.12.0
@@ -33,10 +27,10 @@ dev_dependencies:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0-nullsafety.1
   integration_test:
     path: ../../integration_test
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.4"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
