@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:async';
 
 import 'package:flutter/services.dart';
@@ -43,8 +41,8 @@ void main() {
     };
 
     final List<MethodCall> log = <MethodCall>[];
-    Map<String, dynamic> responses;
-    GoogleSignIn googleSignIn;
+    late Map<String, dynamic> responses;
+    late GoogleSignIn googleSignIn;
 
     setUp(() {
       responses = Map<String, dynamic>.from(kDefaultResponses);
@@ -173,16 +171,16 @@ void main() {
     });
 
     test('concurrent calls of the same method trigger sign in once', () async {
-      final List<Future<GoogleSignInAccount>> futures =
-          <Future<GoogleSignInAccount>>[
+      final List<Future<GoogleSignInAccount?>> futures =
+          <Future<GoogleSignInAccount?>>[
         googleSignIn.signInSilently(),
         googleSignIn.signInSilently(),
       ];
       expect(futures.first, isNot(futures.last),
           reason: 'Must return new Future');
-      final List<GoogleSignInAccount> users = await Future.wait(futures);
+      final List<GoogleSignInAccount?> users = await Future.wait(futures);
       expect(googleSignIn.currentUser, isNotNull);
-      expect(users, <GoogleSignInAccount>[
+      expect(users, <GoogleSignInAccount?>[
         googleSignIn.currentUser,
         googleSignIn.currentUser
       ]);
@@ -218,13 +216,13 @@ void main() {
     });
 
     test('concurrent calls of different signIn methods', () async {
-      final List<Future<GoogleSignInAccount>> futures =
-          <Future<GoogleSignInAccount>>[
+      final List<Future<GoogleSignInAccount?>> futures =
+          <Future<GoogleSignInAccount?>>[
         googleSignIn.signInSilently(),
         googleSignIn.signIn(),
       ];
       expect(futures.first, isNot(futures.last));
-      final List<GoogleSignInAccount> users = await Future.wait(futures);
+      final List<GoogleSignInAccount?> users = await Future.wait(futures);
       expect(
         log,
         <Matcher>[
@@ -248,8 +246,8 @@ void main() {
     });
 
     test('signOut/disconnect methods always trigger native calls', () async {
-      final List<Future<GoogleSignInAccount>> futures =
-          <Future<GoogleSignInAccount>>[
+      final List<Future<GoogleSignInAccount?>> futures =
+          <Future<GoogleSignInAccount?>>[
         googleSignIn.signOut(),
         googleSignIn.signOut(),
         googleSignIn.disconnect(),
@@ -273,8 +271,8 @@ void main() {
     });
 
     test('queue of many concurrent calls', () async {
-      final List<Future<GoogleSignInAccount>> futures =
-          <Future<GoogleSignInAccount>>[
+      final List<Future<GoogleSignInAccount?>> futures =
+          <Future<GoogleSignInAccount?>>[
         googleSignIn.signInSilently(),
         googleSignIn.signOut(),
         googleSignIn.signIn(),
@@ -368,7 +366,7 @@ void main() {
       await googleSignIn.signIn();
       log.clear();
 
-      final GoogleSignInAccount user = googleSignIn.currentUser;
+      final GoogleSignInAccount user = googleSignIn.currentUser!;
       final GoogleSignInAuthentication auth = await user.authentication;
 
       expect(auth.accessToken, '456');
@@ -415,11 +413,11 @@ void main() {
       photoUrl: "https://lh5.googleusercontent.com/photo.jpg",
     );
 
-    GoogleSignIn googleSignIn;
+    late GoogleSignIn googleSignIn;
 
     setUp(() {
       final MethodChannelGoogleSignIn platformInstance =
-          GoogleSignInPlatform.instance;
+          GoogleSignInPlatform.instance as MethodChannelGoogleSignIn;
       platformInstance.channel.setMockMethodCallHandler(
           (FakeSignInBackend()..user = kUserData).handleMethodCall);
       googleSignIn = GoogleSignIn();
@@ -432,7 +430,7 @@ void main() {
     test('can sign in and sign out', () async {
       await googleSignIn.signIn();
 
-      final GoogleSignInAccount user = googleSignIn.currentUser;
+      final GoogleSignInAccount user = googleSignIn.currentUser!;
 
       expect(user.displayName, equals(kUserData.displayName));
       expect(user.email, equals(kUserData.email));
