@@ -29,10 +29,12 @@ const Map<String, dynamic> kDefaultResponses = <String, dynamic>{
   'disconnect': null,
   'isSignedIn': true,
   'getTokens': kTokenData,
+  'requestScopes': true,
 };
 
-final GoogleSignInUserData kUser = getUserDataFromMap(kUserData);
-final GoogleSignInTokenData kToken = getTokenDataFromMap(kTokenData);
+final GoogleSignInUserData? kUser = getUserDataFromMap(kUserData);
+final GoogleSignInTokenData? kToken =
+    getTokenDataFromMap(kTokenData as Map<String, dynamic>);
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -42,7 +44,8 @@ void main() {
     final MethodChannel channel = googleSignIn.channel;
 
     final List<MethodCall> log = <MethodCall>[];
-    Map<String, dynamic> responses; // Some tests mutate some kDefaultResponses
+    late Map<String, dynamic>
+        responses; // Some tests mutate some kDefaultResponses
 
     setUp(() {
       responses = Map<String, dynamic>.from(kDefaultResponses);
