@@ -5,7 +5,7 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart' show required, visibleForTesting;
+import 'package:meta/meta.dart' show visibleForTesting;
 
 import '../google_sign_in_platform_interface.dart';
 import 'types.dart';
@@ -20,11 +20,12 @@ class MethodChannelGoogleSignIn extends GoogleSignInPlatform {
       const MethodChannel('plugins.flutter.io/google_sign_in');
 
   @override
-  Future<void> init(
-      {@required String hostedDomain,
-      List<String> scopes = const <String>[],
-      SignInOption signInOption = SignInOption.standard,
-      String clientId}) {
+  Future<void> init({
+    List<String> scopes = const <String>[],
+    SignInOption signInOption = SignInOption.standard,
+    String? hostedDomain,
+    String? clientId,
+  }) {
     return channel.invokeMethod<void>('init', <String, dynamic>{
       'signInOption': signInOption.toString(),
       'scopes': scopes,
@@ -33,14 +34,14 @@ class MethodChannelGoogleSignIn extends GoogleSignInPlatform {
   }
 
   @override
-  Future<GoogleSignInUserData> signInSilently() {
+  Future<GoogleSignInUserData?> signInSilently() {
     return channel
         .invokeMapMethod<String, dynamic>('signInSilently')
         .then(getUserDataFromMap);
   }
 
   @override
-  Future<GoogleSignInUserData> signIn() {
+  Future<GoogleSignInUserData?> signIn() {
     return channel
         .invokeMapMethod<String, dynamic>('signIn')
         .then(getUserDataFromMap);
@@ -48,12 +49,12 @@ class MethodChannelGoogleSignIn extends GoogleSignInPlatform {
 
   @override
   Future<GoogleSignInTokenData> getTokens(
-      {String email, bool shouldRecoverAuth = true}) {
+      {required String email, bool? shouldRecoverAuth = true}) {
     return channel
         .invokeMapMethod<String, dynamic>('getTokens', <String, dynamic>{
       'email': email,
       'shouldRecoverAuth': shouldRecoverAuth,
-    }).then(getTokenDataFromMap);
+    }).then((result) => getTokenDataFromMap(result!));
   }
 
   @override
@@ -67,23 +68,23 @@ class MethodChannelGoogleSignIn extends GoogleSignInPlatform {
   }
 
   @override
-  Future<bool> isSignedIn() {
-    return channel.invokeMethod<bool>('isSignedIn');
+  Future<bool> isSignedIn() async {
+    return (await channel.invokeMethod<bool>('isSignedIn'))!;
   }
 
   @override
-  Future<void> clearAuthCache({String token}) {
+  Future<void> clearAuthCache({String? token}) {
     return channel.invokeMethod<void>(
       'clearAuthCache',
-      <String, String>{'token': token},
+      <String, String?>{'token': token},
     );
   }
 
   @override
-  Future<bool> requestScopes(List<String> scopes) {
-    return channel.invokeMethod<bool>(
+  Future<bool> requestScopes(List<String> scopes) async {
+    return (await channel.invokeMethod<bool>(
       'requestScopes',
       <String, List<String>>{'scopes': scopes},
-    );
+    ))!;
   }
 }
