@@ -15,7 +15,7 @@ void main() {
     test('Cannot be implemented with `implements`', () {
       expect(() {
         GoogleSignInPlatform.instance = ImplementsGoogleSignInPlatform();
-      }, throwsAssertionError);
+      }, throwsA(isA<Error>()));
     });
 
     test('Can be extended', () {
@@ -23,12 +23,14 @@ void main() {
     });
 
     test('Can be mocked with `implements`', () {
-      final ImplementsGoogleSignInPlatform mock =
-          ImplementsGoogleSignInPlatform();
-      when(mock.isMock).thenReturn(true);
-      GoogleSignInPlatform.instance = mock;
+      GoogleSignInPlatform.instance = ImplementsWithIsMock();
     });
   });
+}
+
+class ImplementsWithIsMock extends Mock implements GoogleSignInPlatform {
+  @override
+  bool get isMock => true;
 }
 
 class ImplementsGoogleSignInPlatform extends Mock
